output
Enter the number of vertices: 4
Enter graph data in matrix form:
6 7 8
 9 7 5
5 4 7
8 7 6
5 7 8 9 8 7
Enter the starting vertex: The node which are reachable are:
BFS is not possible. Not all nodes are reachableBFS is not possible. Not all nodes are reachableBFS is not possible. Not all nodes are reachableBFS is not possible. Not all nodes are reachable
--------------------------------
Process exited after 39.22 seconds with return value 0
Press any key to continue . . .